requirements.txt
flask
RPi.GPIO
flask-socketio
eventlet